+++ fmgen/fmg_wrap.h
@@ -6,7 +6,7 @@
 int OPM_Init(int clock);
 void OPM_Cleanup(void);
 void OPM_Reset(void);
-void OPM_Update(int16_t *buffer, int length, uint8_t *pbsp, uint8_t *pbep);
+void OPM_Update(int16_t *buffer, int length, int16_t *pbsp, int16_t *pbep);
 void FASTCALL OPM_Write(uint32_t r, uint8_t v);
 uint8_t FASTCALL OPM_Read(void);
 void FASTCALL OPM_Timer(uint32_t step);
